Safari-611.1.10/LayoutTests/fast/selectors/querySelector-pseudo-element-inside-functional-pseudo-class-any-expected.txt

This test makes sure that querySelector with pseudo-element inside functional pseudo class :-webkit-any doesn't match anything.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("#test :-webkit-any(p)").length is 1
PASS document.querySelector("#test :-webkit-any(p)") is target
PASS document.querySelectorAll("#test :-webkit-any(p.ng)").length is 1
PASS document.querySelector("#test :-webkit-any(p.ng)") is target
PASS document.querySelectorAll("#test div :-webkit-any(p)").length is 1
PASS document.querySelector("#test div :-webkit-any(p)") is target
PASS document.querySelectorAll("#test div :-webkit-any(p.ng)").length is 1
PASS document.querySelector("#test div :-webkit-any(p.ng)") is target
PASS document.querySelectorAll("#test :-webkit-any(p:first-line)").length is 0
PASS document.querySelector("#test :-webkit-any(p:first-line)") is null
PASS document.querySelectorAll("#test :-webkit-any(div:first-line) p").length is 0
PASS document.querySelector("#test :-webkit-any(div:first-line) p") is null
PASS document.querySelectorAll("#test :-webkit-any(div:first-line) p.ng").length is 0
PASS document.querySelector("#test :-webkit-any(div:first-line) p.ng") is null
PASS successfullyParsed is true

TEST COMPLETE
